requirements.txt
absl-py==0.9.0
actionlib==1.11.13
angles==1.9.11
astor==0.8.1
attrs==19.3.0
bondpy==1.8.3
bz2file==0.98
cachetools==4.0.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.18
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
controller-manager==0.13.3
controller-manager-msgs==0.13.3
controller-manager-tests==0.13.3
cv-bridge==1.12.8
decorator==4.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
dill==0.3.1.1
dopamine-rl==3.0.1
dynamic-reconfigure==1.5.50
Flask==1.1.1
future==0.18.2
gast==0.2.2
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
gevent==1.4.0
gin-config==0.3.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
gym==0.16.0
h5py==2.10.0
httplib2==0.17.0
idna==2.8
image-geometry==1.12.8
imageio==2.6.1
interactive-markers==1.11.4
itsdangerous==1.1.0
Jinja2==2.11.1
jsk-data==2.2.10
jsk-network-tools==2.2.10
jsk-recognition-utils==1.2.9
jsk-tools==2.2.10
jsk-topic-tools==2.2.10
kdl-parser-py==1.12.11
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
laser-geometry==1.6.4
Markdown==3.2.1
MarkupSafe==1.1.1
mesh-tensorflow==0.1.9
message-filters==1.12.14
mpmath==1.1.0
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.32
openni2-launch==0.4.2
opt-einsum==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
pluginlib==1.11.3
promise==2.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.4.10
pypng==0.0.20
python-qt-binding==0.3.4
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.22.0
requests-oauthlib==1.3.0
resource-retriever==1.12.5
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospy==1.12.14
rospy-message-converter==0.5.0
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-joint-trajectory-controller==0.13.5
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rsa==4.0
rviz==1.12.17
scipy==1.4.1
sensor-msgs==1.12.7
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
sound-play==0.3.3
srdfdom==0.4.2
sympy==1.5.1
tensor2tensor==1.15.4
tensorboard==2.1.0
tensorflow-datasets==2.0.0
tensorflow-estimator==2.1.0
tensorflow-gan==2.0.0
tensorflow-gpu==2.1.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.1
tensorflow-probability==0.7.0
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
tqdm==4.42.1
urdfdom-py==0.3.3
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
wiimote==1.13.0
wrapt==1.12.0
xacro==1.11.3
